refactor(textlint): move deprecated code to DEPRECATED/ dir
