+++ packages/textlint/src/DEPRECATED/config/config-loader.ts
