+++ packages/textlint/src/DEPRECATED/config/config-util.ts
