+++ packages/textlint/src/DEPRECATED/config/package-prefix.ts
