+++ packages/textlint/src/DEPRECATED/config/plugin-loader.ts
