+++ packages/textlint/src/DEPRECATED/config/separate-by-config-option.ts
