+++ packages/textlint/src/DEPRECATED/engine/textlint-module-mapper.ts
