+++ packages/textlint/src/DEPRECATED/textlint-package-name-util.ts
